+++ test/ceb.on.spec.js
@@ -58,7 +58,6 @@ describe('ceb.on()', function () {
     context('listen keyboard events', () => {
         let bubblingListener, captureListener, el;
         beforeEach(done => {
-            console.log(typeof KeyboardEvent, navigator.appName, navigator.userAgent);
             bubblingListener = sinon.spy();
             captureListener = sinon.spy();